requirements.txt
tkinter
web3
eth-keys
py-solc-x